Ajoute un texte SEO sur le comparateur de statut
